third_party/blink/web_tests/fast/css/variables/round-trip-inline-style.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<div id='testElem'></div>
<script>

test(function() {
  testElem.style.setProperty('--bar', 'value');
  testElem.setAttribute('style', testElem.getAttribute('style'));
  assert_equals(testElem.style.getPropertyValue('--bar'), 'value');
  testElem.style.setProperty('--bar', ' value');
  testElem.setAttribute('style', testElem.getAttribute('style'));
  assert_equals(testElem.style.getPropertyValue('--bar'), 'value');
}, "round tripping style works for variables.");
</script>
